requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
openai
tiktoken
promptfoo
